+++ scratch.js
@@ -3,18 +3,11 @@ const nlp = require('./src/index')
 // nlp.extend(require('./plugins/numbers/src'))
 // nlp.extend(require('./plugins/dates/src'))
 
-// let doc = nlp('spencer kelly is working here')
-// let res = doc.lookup({
-//   'spencer kelly': 'Cool',
-//   working: 'Uncool',
-// })
-// console.log(res)
+let doc = nlp.tokenize('spencer kelly is working here', {
+  'spencer kelly': 'Person',
+})
 
-let doc = nlp('june the 5th, july the 7th, and sept the 12th.')
-let m = doc.match('[<month>#Month]')
-m.debug()
-m.byName('month').debug()
-
+doc.debug()
 // console.log(m.list[0])
 //by object